\name{statistics}
\alias{statistics}
\docType{data}
\title{
Dataset Containing Answers for the Inducer Term Statistics for the Analysis of The Free Evocation of Words Technique
}
\description{
This dataset is an example created by the developer so that users can have a base as should be the datasets for analysis using this package. All data contained in this dataset were invented and this has no validity on truthfulness.
}
\usage{data("statistics")}
\format{
  A data frame with 50 observations on the following 8 variables.
  \describe{
    \item{\code{IDCLASS}}{a numeric vector}
    \item{\code{WORD01}}{a factor with levels \code{Amplitude} \code{Boxplot} \code{Data} \code{Hypothesis} \code{Inference} \code{Maximum} \code{Mean} \code{Median} \code{Minimum} \code{Mode} \code{Outlier} \code{Percentile} \code{Probability} \code{Quartile} \code{Regression} \code{Sample} \code{Variance}}
    \item{\code{WORD02}}{a factor with levels \code{Amplitude} \code{Bayesian} \code{Boxplot} \code{Data} \code{Expection} \code{Histogram} \code{Inference} \code{Maximum} \code{Mean} \code{Median} \code{Minimum} \code{Mode} \code{Outlier} \code{Parameter} \code{Percentile} \code{Probability} \code{Quartile} \code{Random} \code{Regression} \code{Sample} \code{Variance}}
    \item{\code{WORD03}}{a factor with levels \code{Amplitude} \code{Bayesian} \code{Confidence} \code{Correlation} \code{Data} \code{Estimator} \code{Expection} \code{Histogram} \code{Hypothesis} \code{Inference} \code{Maximum} \code{Mean} \code{Median} \code{Minimum} \code{Outlier} \code{Parameter} \code{Percentile} \code{Probability} \code{Quartile} \code{Random} \code{Regression} \code{Sample} \code{Significance} \code{Variance}}
    \item{\code{WORD04}}{a factor with levels \code{Amplitude} \code{Confidence} \code{Correlation} \code{Data} \code{Estimator} \code{Expection} \code{Hypothesis} \code{Inference} \code{Median} \code{Percentile} \code{Probability} \code{Quartile} \code{Random} \code{Regression} \code{Sample} \code{Significance}}
    \item{\code{WORD05}}{a factor with levels \code{Boxplot} \code{Expection} \code{Hypothesis} \code{Inference} \code{Outlier} \code{Percentile} \code{Probability} \code{Quartile}}
    \item{\code{MEANING}}{a factor with levels \code{A_value_that_flows_out_of_the_standards.} \code{About_fifty_percent_of_the_data.} \code{As_best_represents_Statistics.} \code{As_divided_in_the_middle.} \code{As_important_as_the_Average_and_Median.} \code{Box_plot_split_in_half.} \code{Deduct_about_something_on_the_sample.} \code{Division_of_the_data_into_four_parts.} \code{Division_of_the_data_into_ten_equal_parts.} \code{Find_something_using_Statistics.} \code{Graphic_represented_by_the_Median_and_the_quartiles.} \code{Half_of_the_ordered_values.} \code{I_can_not_explain_very_well_what_it_is.} \code{I_think_it_is_nothing_Fashion.} \code{I_think_it_is_the_sum_of_all_divided_by_something.} \code{I_think_it_is_very_similar_to_the_average.} \code{If_it_is_the_same_for_all_and_so_the_data_is_not_addicted.} \code{In_most_cases_it_is_the_Hope.} \code{It_can_be_arithmetic_or_geometric_or_harmonic.} \code{It_is_a_part_of_the_universe_or_population.} \code{It_is_chance_of_something_occurring.} \code{It_is_equal_to_the_second_quartile.} \code{It_is_the_center_of_the_Histogram.} \code{It_is_the_standard_deviation_squared.} \code{It_is_where_is_concentrated_most.} \code{Main_measure_of_all_distributions.} \code{On_the_order_Percentile_fifty.} \code{One_of_the_most_important_steps_in_tests_Non_parametric.} \code{Putting_it_all_together_and_divide_by_the_total.} \code{Set_events_randomly.} \code{Set_in_a_variable_before_others.} \code{Set_of_elements_we_seek_information.} \code{Set_of_information_about_something.} \code{Set_way_to_find_good_estimators.} \code{Something_very_strong_and_important.} \code{Something_we_want_to_try_to_verify_some_information.} \code{The_biggest_subtracted_from_the_lower_value.} \code{The_data_more_frequently.} \code{The_distance_between_values.} \code{The_distance_of_the_data_in_relation_to_the_average.} \code{The_given_lower_seat.} \code{The_largest_of_all_values.} \code{The_last_value_of_the_data_sorted.} \code{The_least_representative_value_of_the_data.} \code{The_lowest_value_that_appears.} \code{The_maximum_value_of_the_data.} \code{The_minimum_value_of_the_data.} \code{The_point_that_divides_the_bank_in_half.} \code{The_value_that_appears_more_often.} \code{The_value_that_appears_most_of_all.}}
    \item{\code{STRAT}}{a factor with levels \code{Failure} \code{Success}}
  }
}
\examples{
data(statistics)
}
\keyword{datasets}