\name{package}
\alias{package}
\alias{TELP-package}
\docType{package}
\title{
Social Representation Theory Application:
Using The Free Evocation of Words Technique and Other Techniques
}
\value{
  This package does not require any arguments, all analyzes are performed on the provided interface.
  
  Remember that the database must be formatted as follows:
  
  Column 1: Identification code
  
  Column 2: First word
  
  Column 3: Second word
  
  Column 4: Third word
  
  Column 5: Fourth word
  
  Column 6: Fifth word
  
  Column 7: Meaning of the first word
  
  Other columns: Stratification Variables
  
  (*) See Database 'statistics.txt' as an example.
  
}
\description{
Using the TELP method with some functions, this package will make a social representation and other analysis, focusing on the free evocation of words technique with graphics and detailed study of the data.

The Social Representation Theory was developed by social psychologist Serge Moscovici in 1961 with the objective of drawing knowledge before a collective perspective, but without losing the individual perspective. The theory of social representations can be understood by 'private knowledge modality that has the function of the development of behavior and communication between individuals' (Moscovici, 1978). Moreover, this theory has the function to expose individuals to particular subject that may or may not be known, and, in cases of non-knowledge, the purpose of this theory is emerging that the individual subject. The Social Representation can be summarized as the search for the common sense that seeks the view of reality in front of diverse knowledge and even from the perspective of transformation of values and can be divided into two distinct techniques: The Focus Group Technique and The Free Evocation of Words Technique that will be used in this package.

The Free Evocation of Words Technique (TELP, acronym referring Portuguese language) consists of collecting a number of words evoked by a subject facing exposure to an inducer term. The purpose of this technique is to understand the relationships created between words evoked by the individual and the inducer term. This technique is included in the theory of social representations, therefore, on the information transmitted by an individual, seeks to create a profile that define a social group.

The theory of the Central Nucleus can be defined as an expansion of TELP because on this theory, it is defined characterization that will be the basis of social representation developed by TELP.
The theory of the Central Nucleus is based only on the first quadrant of those who are from TELP, and the words sorted in the first quadrant will be used for the preparation of a sentence. This sentence will be one that will characterize the profile of individuals analyzed in relation to the inductive term for which the words were evoked. But the other quadrants also has an important interpretation in different analyzes.
}
\details{
\tabular{ll}{
Package: \tab TELP\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2018-08-31\cr
License: \tab GPL (>2)\cr
}
}
\references{
ABRIC, Jean-Claude. Les representations sociales: aspects theoriques. In: ABRIC, J-C (Org.). Pratiques sociales et representations. Paris: Presses Universitaires de France. 1994. p. 11-37.

ABRIC, Jean-Claude. La recherche du noyau central et de la zone muette des representations sociales. In: ______ (org.). Methodes d'etude des representations sociale. Ramonville Sant-Agne, France: Eres. 2003. p. 58-80.

OLIVEIRA, Denize Cristina de et al. Analise das Evocacoes Livres: uma tecnica de analise estrutural das representacoes sociais. In: MOREIRA, Antonia Silva Paredes; CAMARGO, Brigido Vizeu; JESUINO, Jorge Correia; NOBREGA, Sheva Maia (Org.). Perspectivas Teorico-Metodologicas em Representacoes Sociais. Joao Pessoa: Editora Universitaria UFPB, 2005, p. 573-603.
}
\author{
Gabriel Henrique Oliveira Assuncao

Maintainer: Gabriel Henrique Oliveira Assuncao <gabrielhenryoliver@hotmail.com>
}
\keyword{package}